Wear: Fix issue with invisible settings top/bottom of screen and issue with option selected now showing.
Update various preference screens to display the currently selected value in the summary instead of a static, hardcoded string. This provides better user feedback on the current configuration.

This change affects the following settings:
- Complication tap action
- Heart rate smoothing
- Simplify UI
- Chart timeframe
- Input design

Additionally, add vertical padding to the `ConfigurationActivity` and `WatchfaceConfigurationActivity` to prevent content from being too close to the screen edges.